Initial models for the home feed
A `Reactionable` is any item that a user can like or comment on. I know,
it's kind of (p)refactoring or rather over-engineering at this point.

A `TimelineItem`, the parent class of all items that can land on the
timeline, is a `Reactionable`, and `TextTimelineItem`, a type of
timeline item is a `TimelineItem`.