Preview multiple photos at once
Make sure we're not showing duplicate pictures. This is how our
duplication check works:
Mark files as potential duplicate if they have the same size, type, and
number of bytes. Next, randomly sample bytes from different segments of
the files and compare. If the randomly sampled files are the same then
mark the files as duplicates.